+++ src/main/java/com/sweep/jaksim31/dto/login/KakaoProfile.java
@@ -1,4 +1,4 @@
-package com.sweep.jaksim31.domain.auth;
+package com.sweep.jaksim31.dto.login;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.sweep.jaksim31.dto.login.KakaoLoginRequest;
