+++ src/main/java/com/sweep/jaksim31/service/impl/KaKaoMemberServiceImpl.java
@@ -6,8 +6,8 @@ import com.sweep.jaksim31.auth.TokenProvider;
 import com.sweep.jaksim31.controller.feign.KakaoOAuthInfoFeign;
 import com.sweep.jaksim31.controller.feign.KakaoOAuthLogoutFeign;
 import com.sweep.jaksim31.controller.feign.KakaoOAuthTokenFeign;
-import com.sweep.jaksim31.domain.auth.KakaoOAuth;
-import com.sweep.jaksim31.domain.auth.KakaoProfile;
+import com.sweep.jaksim31.dto.login.KakaoOAuth;
+import com.sweep.jaksim31.dto.login.KakaoProfile;
 import com.sweep.jaksim31.domain.members.MemberRepository;
 import com.sweep.jaksim31.domain.members.Members;
 import com.sweep.jaksim31.domain.token.RefreshToken;
