+++ src/main/java/com/sweep/jaksim31/utils/HeaderUtil.java
@@ -2,6 +2,18 @@ package com.sweep.jaksim31.utils;
 
 import javax.servlet.http.HttpServletRequest;
 
+/**
+ * packageName :  com.sweep.jaksim31.utils
+ * fileName : HeaderUtil
+ * author :  방근호
+ * date : 2023-01-09
+ * description : HeaderUtil 설정 관련 클래스
+ * ===========================================================
+ * DATE                 AUTHOR                NOTE
+ * -----------------------------------------------------------
+ * 2023-01-09           방근호             최초 생성
+ *
+ */
 public class HeaderUtil {
 
     private final static String HEADER_AUTHORIZATION = "Authorization";
